break header out into seperate components
