+++ .data/article.js
@@ -1,2 +1,4 @@
 // @flow
-module.exports = { page: 'Article' };
+const common = require('./common');
+
+module.exports = { ...common, page: 'Article' };
