+++ .data/common.js
@@ -1,0 +1,44 @@
+// @flow
+module.exports.header = {
+    topBarItems: [
+        {
+            text: 'Subscribe',
+            href: '/',
+        },
+        {
+            text: 'Find a job',
+            href: '/',
+        },
+        {
+            text: 'Sign in',
+            href: '/',
+        },
+    ],
+    pillars: [
+        {
+            label: 'News',
+            href: 'http://m.thegulocal.com/uk',
+            pillar: 'news',
+        },
+        {
+            label: 'Opinion',
+            href: 'http://m.thegulocal.com/uk/commentisfree',
+            pillar: 'opinion',
+        },
+        {
+            label: 'Sport',
+            href: 'http://m.thegulocal.com/uk/sport',
+            pillar: 'sport',
+        },
+        {
+            label: 'Culture',
+            href: 'http://m.thegulocal.com/uk/culture',
+            pillar: 'culture',
+        },
+        {
+            label: 'Lifestyle',
+            href: 'http://m.thegulocal.com/uk/lifeandstyle',
+            pillar: 'lifestyle',
+        },
+    ],
+};
