+++ src/components/Header/Nav/Links/Link.js
@@ -1,0 +1,22 @@
+// @flow
+import styled from 'react-emotion';
+
+export default styled('a')({
+    fontSize: 14,
+    fontFamily:
+        '"Guardian Text Sans Web","Helvetica Neue","Helvetica","Arial","Lucida Grande","sans-serif"',
+    color: '#121212',
+    float: 'left',
+    lineHeight: 1.2,
+    position: 'relative',
+    transition: 'color 80ms ease-out',
+    padding: '6px 10px',
+    margin: '1px 0 0',
+    textDecoration: 'none',
+    ':hover': {
+        textDecoration: 'underline',
+    },
+    ':focus': {
+        textDecoration: 'underline',
+    },
+});
